+++ src/pipelayer/__init__.py
@@ -2,7 +2,7 @@
 from __future__ import annotations
 
 from datetime import datetime
-from typing import Any, Callable, List, Optional
+from typing import Any, Callable, List, Optional, Tuple, Type, Union
 
 from pipelayer.context import Context
 from pipelayer.filter import Filter
@@ -12,7 +12,6 @@ from pipelayer.settings import Settings  # NOQA F401
 
 
 class Pipeline:
-
     __name: str
     __context: Context
     __manifest: Manifest
@@ -39,23 +38,27 @@ class Pipeline:
         pipeline.__context = context
         return pipeline
 
-    def run(self, filters: List[Filter], data: Any = None) -> Any:
+    def run(self, filters: List[Union[Filter, Callable[[Context, Any], Any]]], data: Any = None) -> Any:
         """
         The Pipeline runner
         """
         self.__manifest = Manifest(name=self.name, start=datetime.utcnow())
 
         for filter in filters:
-            manifest_entry = FilterManifestEntry(name=filter.name, start=datetime.utcnow())
+            filter_name = filter.name if isinstance(filter, Filter) else filter.__name__
+            manifest_entry = FilterManifestEntry(name=filter_name, start=datetime.utcnow())
+
+            # initialize
+            filter_func, pre_process, post_process = self.__initialize_filter(filter)
 
             # pre_process
-            data = Pipeline.__run_filter_process(self.context, manifest_entry, filter.pre_process, data)
+            manifest_entry.pre_process, data = self.__run_filter_process(pre_process, data)
 
             # filter
-            data = filter.run(self.context, data)
+            data = filter_func(self.context, data)
 
             # post_process
-            data = Pipeline.__run_filter_process(self.context, manifest_entry, filter.post_process, data, False)
+            manifest_entry.post_process, data = self.__run_filter_process(post_process, data)
 
             manifest_entry.end = datetime.utcnow()
             manifest_entry.duration = (manifest_entry.end - manifest_entry.start)
@@ -69,8 +72,22 @@ class Pipeline:
         return data
 
     @staticmethod
-    def __run_filter_process(context: Context, filter_manifest_entry: FilterManifestEntry,
-                             process: Optional[Callable], data: Any, pre_process: bool = True) -> Any:
+    def __initialize_filter(
+        filter: Union[Filter, Type[Filter], Callable[[Context, Any], Any]]
+    ) -> Tuple[Callable[[Context, Any], Any], Union[Callable[..., Any], None], Union[Callable[..., Any], None]]:
+
+        if type(filter) is Filter and not isinstance(filter, Filter):
+            # The checks should have isolated the type, but mypy complains
+            filter = filter()  # type: ignore
+
+        if isinstance(filter, Filter):
+            return filter.run, filter.pre_process, filter.post_process
+
+        func = filter if type(filter) == callable else lambda context, data: data
+
+        return func, None, None
+
+    def __run_filter_process(self, process: Optional[Callable], data: Any) -> Tuple[ManifestEntry, Any]:
         """
         The filter pre/post process runner
         """
@@ -79,15 +96,11 @@ class Pipeline:
 
         process_manifest_entry = ManifestEntry(name=process.__name__, start=datetime.utcnow())
 
-        data = process(context, data)
+        data = process(self.context, data)
 
         process_manifest_entry.end = datetime.utcnow()
         process_manifest_entry.duration = (process_manifest_entry.end - process_manifest_entry.start)
 
-        if pre_process:
-            filter_manifest_entry.pre_process = process_manifest_entry
-        else:
-            filter_manifest_entry.post_process = process_manifest_entry
-        return data
+        return process_manifest_entry, data
 
     __metaclass__ = Final