+++ HISTORY.md
@@ -1,2 +1,6 @@
+## 0.2.0
+* Allow for static/module/lamba methods to be used as Filters
+* Allow type(filter) to be added to Pipeline.filters
+
 ## 0.1.0
 Initial release.